backend/requirements.txt
fastapi[standard]
passlib[bcrypt]
sqlmodel
psycopg[binary]
pydantic_settings 
psycopg2